pyproject.toml
[project]
name = "extract-metadata-from-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pillow-heif>=1.3.0",
    "python-docx>=1.1.0",
]
